# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KABBAK API",
    "version": "0.2.0",
    "description": "Read-only API backend for the KABBAK frontend. All content responses use a {data, meta?} envelope (see README). Protected routes accept x-api-key or Authorization: Bearer <key>. Named clients can carry a basic, premium, or pro+ accessLevel, plus optional roles and scopes for privileged admin routes.",
    "x-service-name": "kabbak-api"
  },
  "servers": [
    {
      "url": "http://localhost:3100/api/v1",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/tarot/cards": {
      "get": {
        "tags": [
          "tarot"
        ],
        "summary": "List tarot cards",
        "operationId": "listTarotCards",
        "parameters": [
          {
            "$ref": "#/components/parameters/SearchQuery"
          },
          {
            "$ref": "#/components/parameters/ArcanaFilter"
          },
          {
            "$ref": "#/components/parameters/SuitFilter"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/tarot/cards/{cardId}": {
      "get": {
        "tags": [
          "tarot"
        ],
        "summary": "Get one tarot card",
        "operationId": "getTarotCard",
        "parameters": [
          {
            "$ref": "#/components/parameters/CardId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/tattvas": {
      "get": {
        "tags": ["tattvas"],
        "summary": "List tattva reference entries",
        "operationId": "listTattvas",
        "responses": {
          "200": { "$ref": "#/components/responses/JsonPayload" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/tattvas/{tattvaId}": {
      "get": {
        "tags": ["tattvas"],
        "summary": "Get one tattva reference entry",
        "operationId": "getTattva",
        "parameters": [
          { "name": "tattvaId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/JsonPayload" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/kabbalah/tree": {
      "get": {
        "tags": [
          "kabbalah"
        ],
        "summary": "Get the Kabbalah tree dataset",
        "operationId": "getKabbalahTree",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/kabbalah/sephiroth/{value}": {
      "get": {
        "tags": [
          "kabbalah"
        ],
        "summary": "Get one sephirah entry",
        "operationId": "getSephirah",
        "parameters": [
          {
            "$ref": "#/components/parameters/SephirahValue"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/kabbalah/paths/{value}": {
      "get": {
        "tags": [
          "kabbalah"
        ],
        "summary": "Get one Kabbalah path entry",
        "operationId": "getKabbalahPath",
        "parameters": [
          {
            "$ref": "#/components/parameters/KabbalahPathValue"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/kabbalah/cube": {
      "get": {
        "tags": [
          "kabbalah"
        ],
        "summary": "Get the Kabbalah cube dataset",
        "operationId": "getKabbalahCube",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/numbers": {
      "get": {
        "tags": [
          "numbers"
        ],
        "summary": "List number datasets",
        "operationId": "listNumbers",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/numbers/{value}": {
      "get": {
        "tags": [
          "numbers"
        ],
        "summary": "Get one number entry",
        "operationId": "getNumber",
        "parameters": [
          {
            "$ref": "#/components/parameters/NumberValue"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/playing-cards": {
      "get": {
        "tags": [
          "playing-cards"
        ],
        "summary": "List playing card datasets",
        "operationId": "listPlayingCards",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/playing-cards/{cardId}": {
      "get": {
        "tags": [
          "playing-cards"
        ],
        "summary": "Get one playing card entry or alias match",
        "operationId": "getPlayingCard",
        "parameters": [
          {
            "$ref": "#/components/parameters/PlayingCardId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonPayload"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "ResponseMeta": {
        "type": "object",
        "description": "Optional metadata attached to success responses for pagination, tracing and versioning.",
        "properties": {
          "requestId": {
            "type": "string",
            "description": "Unique identifier for the request, echoed from x-request-id or generated."
          },
          "version": {
            "type": "string",
            "description": "API/service version at the time of the response. Allows GUIs to detect shape changes."
          },
          "total": {
            "type": "integer",
            "description": "Total number of items available (for list/paginated responses)."
          },
          "offset": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "count": {
            "type": "integer",
            "description": "Number of items in this response page."
          }
        }
      },
      "ApiSuccess": {
        "type": "object",
        "description": "Standard success envelope for all content API responses. GUIs should expect data + optional meta.",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "description": "Primary payload. May be object, array, or primitive depending on endpoint."
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "additionalProperties": false
      },
      "ApiError": {
        "type": "object",
        "required": [
          "error",
          "message",
          "requestId"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "description": "Optional structured error details.",
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      },
...
    },
    "responses": {
      "JsonPayload": {
        "description": "Standard JSON success response using the ApiSuccess envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiSuccess"
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "A valid API key was not supplied.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "The requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
    },
    "parameters": {
...
      "CardId": {
        "name": "cardId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "SearchQuery": {
        "name": "q",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "ArcanaFilter": {
        "name": "arcana",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "SuitFilter": {
        "name": "suit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
...
      "SephirahValue": {
        "name": "value",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "KabbalahPathValue": {
        "name": "value",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "NumberValue": {
        "name": "value",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
...
      "PlayingCardId": {
        "name": "cardId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    }
  }
}